requirements.txt
importlib-metadata==6.8.0
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
Pygments==2.16.1
pyserial==3.5
rich==13.6.0
textual==0.40.0
typing_extensions==4.8.0
uc-micro-py==1.0.2
zipp==3.17.0
